+++ Source/equipment/_equipmentDictionary.js
@@ -106,6 +106,7 @@ for (const file of [
 	"warcry-tormenting.js",
 	"warhammer-base.js",
 	"warhammer-piercing.js",
+	"warhammer-reactive.js",
 	"warhammer-slowing.js"
 ]) {
 	const equip = require(`./${file}`);