+++ Source/equipment/warhammer-base.js
@@ -2,10 +2,10 @@ const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { dealDamage, addModifier } = require('../combatantDAO.js');
 const { needsLivingTargets } = require('../enemyDAO.js');
 
-module.exports = new EquipmentTemplate("Warhammer", "Strike a foe for @{damage} (+@{bonus} if foe is already stunned) @{element} damage", "Damage x@{critBonus}", "Earth", needsLivingTargets(effect))
+module.exports = new EquipmentTemplate("Warhammer", "Strike a foe for @{damage} (+@{bonus} if foe is currently stunned) @{element} damage", "Damage x@{critBonus}", "Earth", needsLivingTargets(effect))
 	.setCategory("Weapon")
 	.setTargetingTags({ target: "single", team: "enemy" })
-	.setUpgrades("Piercing Warhammer", "Slowing Warhammer")
+	.setUpgrades("Piercing Warhammer", "Reactive Warhammer", "Slowing Warhammer")
 	.setModifiers([{ name: "Stagger", stacks: 1 }])
 	.setCost(350)
 	.setUses(15)